Wrapping text in a div tag
When I opened the live new-page.html today on my home Linux machine using Firefox in dark mode, I couldn't see the text.  I think it was black text on a black background.  When I selected everything on the page, I could see the text.

I think wrapping the text inside a div with the "main" class will make the text visible in a lighter background.